move water options to item config
